changed double to int

+++ src/common/SignalTools.java
@@ -64,7 +64,7 @@ public class SignalTools {
 		return tab;
 	}
 	
-	public static double getWartoscSrednia(double[][] sygnal, double f, double d) {
+	public static double getWartoscSrednia(double[][] sygnal, int f, double d) {
 		double suma = 0;
 		for(int n = 0; n < sygnal.length; n++) {
 			suma += sygnal[n][1];
@@ -77,7 +77,7 @@ public class SignalTools {
 		return srednia;
 	}
 	
-	public static double getWartoscSredniaBezwzgledna(double[][] sygnal, double f, double d) {
+	public static double getWartoscSredniaBezwzgledna(double[][] sygnal, int f, double d) {
 		double suma = 0;
 		for(int n = 0; n < sygnal.length; n++) {
 			suma += Math.abs(sygnal[n][1]);
@@ -89,7 +89,7 @@ public class SignalTools {
 		return srednia;
 	}
 	
-	public static double getWartoscSredniaMoc(double[][] sygnal, double f, double d) {
+	public static double getWartoscSredniaMoc(double[][] sygnal, int f, double d) {
 		double suma = 0;
 		for(int n = 0; n < sygnal.length; n++) {
 			suma += Math.pow((sygnal[n][1]), 2);
@@ -101,7 +101,7 @@ public class SignalTools {
 		return srednia;
 	}
 	
-	public static double getWartoscWariancja(double[][] sygnal, double f, double d, double sredniaSygnalu) {
+	public static double getWartoscWariancja(double[][] sygnal, int f, double d, double sredniaSygnalu) {
 		double suma = 0;
 		for(int n = 0; n < sygnal.length; n++) {
 			suma += Math.pow((sygnal[n][1] - sredniaSygnalu), 2);
